Move resolve_service to its own file
This is useful outside endpoints.
